packages/email_support/send_newsletter/requirements.txt
requests==2.31.0
logtail-python==0.2.10